Refactorització de rating de Double a Float

+++ app/src/main/java/com/abeltran10/carajilloapp/data/repo/BarRepository.java
@@ -74,7 +74,7 @@ public class BarRepository {
                 bar.setAddress(address + " " + number);
                 bar.setCity(city);
                 bar.setPostalCode(postalCode);
-                bar.setRating((Double) map.get("rating"));
+                bar.setRating((Float) map.get("rating"));
 
                 setBar(bar);
